+++ packages/contracts/eslint.config.js
@@ -1,0 +1,3 @@
+import { typescriptBaseConfig } from "@planning-poker/config/eslint/base";
+
+export default [...typescriptBaseConfig, { ignores: ["dist/**"] }];