+++ planning-poker-fe/src/shared/types/index.ts
@@ -1,3 +1,0 @@
-export type { Room } from "./room.model";
-export type { Participant } from "./participant.model";
-export type { ValueSet } from "./value-set.model";